+++ src/processImageHelpers.js
@@ -10,8 +10,10 @@ export function extractFileName(response) {
     return matches && matches[1] ? matches[1] : console.error("No matches found", matches);
 }
 
-export function buildImageFileName(imageName, fileType) {
-    const outputFileName = `${imageName}.${fileType.ext}`;
-    const destinationFilePath = path.join(process.env.IMAGE_INPUT_PATH, outputFileName);
-    return destinationFilePath;
+export function buildImageFileNameAndPath(imageName, fileType) {
+    return path.join(process.env.IMAGE_INPUT_PATH, `${imageName}.${fileType.ext}`);
+}
+
+async function getFileType(response) {
+    const fileType = await fileTypeFromFile(response)
 }
